feet: 피드백리뷰 엔티티 Rating 타입 수정

+++ src/main/java/com/example/feedprep/domain/feedbackreview/entity/FeedbackReview.java
@@ -35,7 +35,7 @@ public class FeedbackReview  extends  BaseTimeEntity{
 
 		private Long tutorId;
 
-		private int rating;
+		private Long rating;
 
 		private  String content;
